+++ packages/agentmark/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @puzzlet/agentmark
 
+## 2.1.3
+
+### Patch Changes
+
+- 27a4e17: Adding proper types for output
+
+## 2.1.2
+
+### Patch Changes
+
+- a93359d: Chore: add output type
+
+## 2.1.1
+
+### Patch Changes
+
+- b1660e9: Fix to add default/fallback types for file loader
+
+## 2.1.0
+
+### Minor Changes
+
+- 7ba382b: Feat: Add Template Runner
+
+## 2.0.0
+
+### Major Changes
+
+- adab584: Type safety. Single returned result.
+
 ## 1.2.2
 
 ### Patch Changes